refactor: extract incompatible patch stuff to an action
