+++ app/Connect/Actions/BuildClientPatchDataAction.php
@@ -12,7 +12,6 @@ use App\Models\GameHash;
 use App\Models\PlayerGame;
 use App\Models\User;
 use App\Platform\Enums\AchievementFlag;
-use App\Platform\Services\VirtualGameIdService;
 use Illuminate\Database\Eloquent\Collection;
 use InvalidArgumentException;
 
@@ -55,7 +54,10 @@ class BuildClientPatchDataAction
             $gameHash->compatibility !== GameHashCompatibility::Compatible
             && $gameHash->compatibility_tester_id !== $user?->id
         ) {
-            return $this->buildIncompatiblePatchData($game ?? $gameHash->game, $gameHash->compatibility);
+            return (new BuildIncompatiblePatchDataAction())->execute(
+                $game ?? $gameHash->game,
+                $gameHash->compatibility
+            );
         }
 
         $rootGameId = (new ResolveRootGameIdFromGameAndGameHashAction())->execute($gameHash, $game, $user);
@@ -223,37 +225,5 @@ class BuildClientPatchDataAction
         }
 
         return max(1, $gamePlayerCount);
-    }
-
-    /**
-     * @param Game $game The game to build root-level data for
-     */
-    private function buildIncompatiblePatchData(
-        Game $game,
-        GameHashCompatibility $gameHashCompatibility,
-    ): array {
-        $seeSupportedGameFiles = 'See the Supported Game Files page for this game to find a compatible version.';
-
-        return [
-            'Success' => true,
-            'PatchData' => [
-                'ID' => VirtualGameIdService::encodeVirtualGameId($game->id, $gameHashCompatibility),
-                'Title' => 'Unsupported Game Version',
-                'ConsoleID' => $game->ConsoleID,
-                'ImageIcon' => $game->ImageIcon,
-                'ImageIconURL' => media_asset($game->ImageIcon),
-                'Achievements' => [
-                    (new CreateWarningAchievementAction())->execute(
-                        title: 'Unsupported Game Version',
-                        description: match ($gameHashCompatibility) {
-                            GameHashCompatibility::Incompatible => "This version of the game is known to not work with the defined achievements. $seeSupportedGameFiles",
-                            GameHashCompatibility::Untested => "This version of the game has not been tested to see if it works with the defined achievements. $seeSupportedGameFiles",
-                            GameHashCompatibility::PatchRequired => "This version of the game requires a patch to support achievements. $seeSupportedGameFiles",
-                            default => $seeSupportedGameFiles,
-                        }),
-                ],
-                'Leaderboards' => [],
-            ],
-        ];
     }
 }
